requirements.txt
Brotli @ file:///croot/brotli-split_1714483155106/work
certifi @ file:///croot/certifi_1725551672989/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
coloredlogs==15.0.1
contourpy @ file:///croot/contourpy_1700583582875/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
filelock @ file:///croot/filelock_1700591183607/work
flatbuffers==24.3.25
fonttools @ file:///croot/fonttools_1713551344105/work
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645438755360/work
humanfriendly==10.0
idna @ file:///croot/idna_1714398848350/work
importlib_resources @ file:///croot/importlib_resources-suite_1720641103994/work
Jinja2 @ file:///croot/jinja2_1730902924303/work
joblib @ file:///croot/joblib_1718217211762/work
kiwisolver @ file:///croot/kiwisolver_1672387140495/work
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
matplotlib==3.9.2
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
mpmath @ file:///croot/mpmath_1690848262763/work
networkx @ file:///croot/networkx_1717597493534/work
numpy @ file:///croot/numpy_and_numpy_base_1708638617955/work/dist/numpy-1.26.4-cp39-cp39-linux_x86_64.whl#sha256=6094eeedd869502faa0fd0a8c5ad3a70c5779be06ddd1feb7627e5c212fac420
onnx==1.17.0
onnxruntime-gpu==1.19.2
packaging @ file:///croot/packaging_1720101850331/work
pandas @ file:///home/conda/feedstock_root/build_artifacts/pandas_1726878406085/work
pillow @ file:///croot/pillow_1731594689713/work
ply==3.11
protobuf==5.29.1
pyparsing @ file:///croot/pyparsing_1731445506121/work
pyprojroot @ file:///home/conda/feedstock_root/build_artifacts/pyprojroot_1679412682637/work
PyQt5==5.15.10
PyQt5-sip @ file:///croot/pyqt-split_1698769088074/work/pyqt_sip
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1731919281354/work
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1706886791323/work
PyYAML @ file:///croot/pyyaml_1728657952215/work
requests @ file:///croot/requests_1730999120400/work
scikit-learn @ file:///croot/scikit-learn_1721921875708/work
scipy @ file:///croot/scipy_1717521478074/work/dist/scipy-1.13.1-cp39-cp39-linux_x86_64.whl#sha256=2cb3d1436f8762a53161486cb9493ccc305fc760e8f830f82888792cbaf99099
sip @ file:///croot/sip_1698675935381/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sympy @ file:///croot/sympy_1724938189289/work
threadpoolctl @ file:///croot/threadpoolctl_1719407800858/work
tomli @ file:///opt/conda/conda-bld/tomli_1657175507142/work
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado @ file:///croot/tornado_1718740109488/work
tqdm @ file:///croot/tqdm_1724853939799/work
triton==3.1.0
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1727140567071/work
unicodedata2 @ file:///croot/unicodedata2_1713212950228/work
urllib3 @ file:///croot/urllib3_1727769808118/work
zipp @ file:///croot/zipp_1732630741423/work
